Only overlay if the pfPort is actually a port, and not a random directory that happens to share the name with a port...

+++ builder_scripts/rebuild_package_binaries_pbi.php
@@ -193,7 +193,7 @@ function overlay_pfPort($port_path) {
 	// If a pfport exists, overlay that folder onto $port_path
 	$port_name = basename($port_path);
 	$port_path = "/usr/ports/" . $port_path;
-	if (file_exists("{$pfports}/{$port_name}") && is_dir("{$pfports}/{$port_name}")) {
+	if (file_exists("{$pfports}/{$port_name}") && is_dir("{$pfports}/{$port_name}") && is_file("{$pfports}/{$port_name}/Makefile")) {
 		echo ">>> [" . date("H:i:s") . "] Overelaying pfPort {$port_name} onto {$port_path} ... ";
 		if (file_exists($port_path) && is_dir($port_path)) {
 			system("/bin/rm -rf {$port_path}.orig");
